[IR] Workaround for possible crashes during deserialization from zipped klibs
^KT-73826

+++ compiler/util-klib/src/org/jetbrains/kotlin/library/impl/KotlinLibraryLayoutImpl.kt
@@ -103,10 +103,16 @@ class FromZipIrLibraryImpl(zipped: IrLibraryLayoutImpl, zipFileSystem: FileSyste
 fun KotlinLibraryLayoutImpl.extract(file: File): File = extract(this.klib, file)
 
 private fun extract(zipFile: File, file: File) = zipFile.withZipFileSystem { zipFileSystem ->
-    val temporary = org.jetbrains.kotlin.konan.file.createTempFile(file.name)
-    zipFileSystem.file(file).copyTo(temporary)
-    temporary.deleteOnExit()
-    temporary
+    val innerFile = zipFileSystem.file(file)
+    if (innerFile.exists) {
+        val temporary = org.jetbrains.kotlin.konan.file.createTempFile(file.name)
+        innerFile.copyTo(temporary)
+        temporary.deleteOnExit()
+        temporary
+    } else {
+        // return deliberately nonexistent file name zipFile.name + ! + file.name
+        File(zipFile.path + "!" + file.path)
+    }
 }
 
 fun KotlinLibraryLayoutImpl.extractDir(directory: File): File = extractDir(this.klib, directory)
